requirements.txt
django==3.2.5
djangorestframework==3.12.4 
mysqlclient==2.0.3
attrs==21.2.0
cattrs==1.8.0
dependency-injector==4.35.2
drf_yasg==1.20.0
django-environ==0.5.0
python-dateutil==2.8.2
django-cors-headers==3.8.0
pillow==8.3.2